+++ CHANGELOG.md
@@ -7,6 +7,9 @@
 ### **Copyrigth**: 
 *Tiledesk SRL*
 
+# this branch
+- **bug-fix**: Updated MCP tools in the ai-prompt action when another ai-prompt action is selected
+
 
 # 1.39.9-rc12:
 - **bug-fix**: renamed "selected Mcp Servers"